packages/dbt-adapter/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "usf-dbt-adapter"
version = "0.1.0"
description = "Import dbt Semantic Layer metrics into USF SDL format"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6",
    "pydantic>=2.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-cov"]
